requirements.txt
board==1.0
paho-mqtt==2.1.0
pytz==2024.1
RPi.bme280==0.2.4
smbus2==0.4.3